+++ src/view/InputView.js
@@ -1,4 +1,6 @@
+// constants
 import { MESSAGE } from '../constants/System.js';
+// utils
 import readLineAsync from '../utils/ReadLine.js';
 
 const InputView = {
